Melhorada orientação a objetos
