+++ src/java/br/ufscar/dc/dsw/dao/DAOTeatro.java
@@ -11,18 +11,10 @@
 //import java.sql.Date;
 //import java.util.List;
 //
-//public class DAOTeatro {
+//public class DAOTeatro extends DBConnection {
 //
 //    public DAOTeatro() {
-//        try {
-//            Class.forName("org.apache.derby.jdbc.ClientDriver");
-//        } catch (ClassNotFoundException e) {
-//            throw new RuntimeException(e);
-//        }
-//    }
-//
-//    protected Connection getConnection() throws SQLException {
-//        return DriverManager.getConnection("jdbc:derby://localhost:1527/dsw_1", "root", "root");
+//        super();
 //    }
 //
 //    public void insert(Usuario usuario) {
