+++ src/java/br/ufscar/dc/dsw/dao/DAOUsuario.java
@@ -12,18 +12,10 @@ import java.util.ArrayList;
 import java.sql.Date;
 import java.util.List;
 
-public class DAOUsuario {
+public class DAOUsuario extends DBConnection {
 
     public DAOUsuario() {
-        try {
-            Class.forName("org.apache.derby.jdbc.ClientDriver");
-        } catch (ClassNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    protected Connection getConnection() throws SQLException {
-        return DriverManager.getConnection("jdbc:derby://localhost:1527/dsw_1", "root", "root");
+        super();
     }
 
     public void insert(Usuario usuario) {
