+++ src/java/br/ufscar/dc/dsw/dao/DBConnection.java
@@ -1,0 +1,20 @@
+package br.ufscar.dc.dsw.dao;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class DBConnection {
+
+    public DBConnection() {
+        try {
+            Class.forName("org.apache.derby.jdbc.ClientDriver");
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
+    }
+    
+    protected Connection getConnection() throws SQLException {
+        return DriverManager.getConnection("jdbc:derby://localhost:1527/dsw_1", "root", "root");
+    }
+}
